Cooper/security update, adds many security features and should not impede any features

+++ nextstep/src/main/java/com/NextStep/nextstep/repository/UserAccountRepository.java
@@ -1,10 +1,9 @@
 package com.NextStep.nextstep.repository;
 
-import com.NextStep.nextstep.entity.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.NextStep.nextstep.entity.UserAccount;
 
-// Adding <UserAccount, Long> fixes the "incompatible types" error
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer>{
+
     UserAccount findByEmail(String email);
 }
-
